Sim: Add "number of bits" parameter to KIBIS PRBS model and some more fixes
The other fixes:

- PRBS's `m_bits` should be an integer.
- Fix some references to outdated parameter names.

+++ eeschema/sim/kibis/kibis.h
@@ -113,7 +113,7 @@ public:
     };
     double m_bitrate = 10e6;
     double m_delay = 0;
-    double m_bits = 10;
+    int m_bits = 10;
 
     std::vector<std::pair<int, double>> GenerateBitSequence() override;
     bool Check( IbisWaveform* aRisingWf, IbisWaveform* aFallingWf ) override;
